+++ index.php
@@ -1,0 +1,23 @@
+<html lang='pt-BR'>
+	<head>
+		<title>Pages that you can see</title>
+	</head>
+	<body>
+		
+		<h2>Pages that you can see</h2>
+		
+		<ol class=''>
+			<li>
+				<a href='test/httpError_test.php'>test/httpError_test.php</a>
+			</li>
+			<li>
+				<a href='test/Person_test.php'>test/Person_test.php</a>
+			</li>
+			<li>
+				<a href='pages/person.php'>pages/person.php</a>
+			</li>
+		</ol>
+		
+		
+	</body>
+</html>